+++ rpc/tester/tester_test.go
@@ -31,10 +31,10 @@ const (
 var addr = fmt.Sprintf("http://%s:%d", host, port)
 
 type Request struct {
-	Version string   `json:"jsonrpc"`
-	Method  string   `json:"method"`
-	Params  []string `json:"params"`
-	ID      int      `json:"id"`
+	Version string      `json:"jsonrpc"`
+	Method  string      `json:"method"`
+	Params  interface{} `json:"params"`
+	ID      int         `json:"id"`
 }
 
 type RPCError struct {
@@ -49,7 +49,7 @@ type Response struct {
 	Result json.RawMessage `json:"result,omitempty"`
 }
 
-func createRequest(method string, params []string) Request {
+func createRequest(method string, params interface{}) Request {
 	return Request{
 		Version: "2.0",
 		Method:  method,
@@ -58,7 +58,7 @@ func createRequest(method string, params []string) Request {
 	}
 }
 
-func call(method string, params []string) (*Response, error) {
+func call(t *testing.T, method string, params interface{}) (*Response, error) {
 	req, err := json.Marshal(createRequest(method, params))
 	if err != nil {
 		return nil, err
@@ -67,23 +67,23 @@ func call(method string, params []string) (*Response, error) {
 	/* #nosec */
 	res, err := http.Post(addr, "application/json", bytes.NewBuffer(req))
 	if err != nil {
-		return nil, err
+		t.Fatal(err)
 	}
 
 	decoder := json.NewDecoder(res.Body)
 	var rpcRes *Response
 	err = decoder.Decode(&rpcRes)
 	if err != nil {
-		return nil, err
+		t.Fatal(err)
 	}
 
 	if rpcRes.Error != nil {
-		return nil, errors.New(rpcRes.Error.Message)
+		t.Fatal(errors.New(rpcRes.Error.Message))
 	}
 
 	err = res.Body.Close()
 	if err != nil {
-		return nil, err
+		t.Fatal(err)
 	}
 
 	return rpcRes, nil
@@ -92,7 +92,7 @@ func call(method string, params []string) (*Response, error) {
 func TestEth_protocolVersion(t *testing.T) {
 	expectedRes := hexutil.Uint(version.ProtocolVersion)
 
-	rpcRes, err := call("eth_protocolVersion", []string{})
+	rpcRes, err := call(t, "eth_protocolVersion", []string{})
 	require.NoError(t, err)
 
 	var res hexutil.Uint
@@ -104,7 +104,7 @@ func TestEth_protocolVersion(t *testing.T) {
 }
 
 func TestEth_blockNumber(t *testing.T) {
-	rpcRes, err := call("eth_blockNumber", []string{})
+	rpcRes, err := call(t, "eth_blockNumber", []string{})
 	require.NoError(t, err)
 
 	var res hexutil.Uint64
@@ -115,7 +115,7 @@ func TestEth_blockNumber(t *testing.T) {
 }
 
 func TestEth_GetBalance(t *testing.T) {
-	rpcRes, err := call("eth_getBalance", []string{addrA, "0x0"})
+	rpcRes, err := call(t, "eth_getBalance", []string{addrA, "0x0"})
 	require.NoError(t, err)
 
 	var res hexutil.Big
@@ -132,7 +132,7 @@ func TestEth_GetBalance(t *testing.T) {
 
 func TestEth_GetStorageAt(t *testing.T) {
 	expectedRes := hexutil.Bytes{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
-	rpcRes, err := call("eth_getStorageAt", []string{addrA, string(addrAStoreKey), "0x0"})
+	rpcRes, err := call(t, "eth_getStorageAt", []string{addrA, string(addrAStoreKey), "0x0"})
 	require.NoError(t, err)
 
 	var storage hexutil.Bytes
@@ -146,7 +146,7 @@ func TestEth_GetStorageAt(t *testing.T) {
 
 func TestEth_GetCode(t *testing.T) {
 	expectedRes := hexutil.Bytes{}
-	rpcRes, err := call("eth_getCode", []string{addrA, "0x0"})
+	rpcRes, err := call(t, "eth_getCode", []string{addrA, "0x0"})
 	require.NoError(t, err)
 
 	var code hexutil.Bytes
@@ -156,4 +156,16 @@ func TestEth_GetCode(t *testing.T) {
 
 	t.Logf("Got code [%X] for %s\n", code, addrA)
 	require.True(t, bytes.Equal(expectedRes, code), "expected: %X got: %X", expectedRes, code)
+}
+
+func TestEth_NewFilter(t *testing.T) {
+	param := make([]map[string][]string, 1)
+	param[0] = make(map[string][]string)
+	param[0]["topics"] = []string{"0x0000000000000000000000000000000000000000000000000000000012341234"}
+	rpcRes, err := call(t, "eth_newFilter", param)
+	require.NoError(t, err)
+
+	var code hexutil.Bytes
+	err = code.UnmarshalJSON(rpcRes.Result)
+	require.NoError(t, err)
 }