+++ x/evm/client/cli/tx.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/context"
+	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth"
@@ -33,7 +34,7 @@ func GetTxCmd(storeKey string, cdc *codec.Codec) *cobra.Command {
 		RunE:                       client.ValidateCmd,
 	}
 
-	evmTxCmd.AddCommand(client.PostCommands(
+	evmTxCmd.AddCommand(flags.PostCommands(
 		GetCmdGenTx(cdc),
 		GetCmdGenCreateTx(cdc),
 	)...)
@@ -73,7 +74,7 @@ func GetCmdGenTx(cdc *codec.Codec) *cobra.Command {
 
 				data, err = hexutil.Decode(payload)
 				if err != nil {
-					fmt.Println(err)
+					return err
 				}
 			}
 
@@ -117,7 +118,7 @@ func GetCmdGenCreateTx(cdc *codec.Codec) *cobra.Command {
 
 			data, err := hexutil.Decode(payload)
 			if err != nil {
-				fmt.Println(err)
+				return err
 			}
 
 			var amount int64